Raise exception if upgrade not implemented
The function will only be called if the __init__ file indicates that the plug-in can upgrade the configuration type. If the __init__ file indicates it can upgrade, but the upgrade function itself is not implemented, this commit makes it so that an exception is raised. This exception is then caught by the version upgrade manager with an appropriate message (after which it will continue with the next file).

Contributes to issue CURA-844.