Add information how to store new config files
To save a configuration file, the version upgrade system needs to know where to save it now with the new version, and what extension to give it. The extension is obtained from the MIME type database, so instead of the extension we give it the MIME type.

Contributes to issue CURA-844.